Fixed permissions in qr and enabled to work with python 3 (as well as 2).
Former-commit-id: 56caf6dd9e3860f96061ac44d42e981288730903

+++ python/examples/qr/src/qr.py
@@ -227,7 +227,7 @@ def joinMatrix(A):
 
 def splitMatrix(A, MSIZE):
     splittedMatrix = []
-    bSize = len(A) / MSIZE
+    bSize = int(len(A) / MSIZE)
     for i in range(MSIZE):
         splittedMatrix.append([])
         for j in range(MSIZE):
